fix: place refresh below the header
